+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Rendered Markdown links now retain the outer formatting (e.g., if you wrapped the whole link inside bold or emphasis).
 - The Zettlr default `tex`-template now doesn't break checkbox exports by including the `amsmath` and `amssymb`-packages.
 - New shortcuts: Pressing `Ctrl+Enter` will insert a new line below the one in which you are currently, placing your cursor there. Pressing `Ctrl+Shift+Enter` will do the same but above the current line.
+- Added context menu entries to copy mail addresses and links to clipboard.
 
 ## Under the Hood